## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [ ] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [ ] Я запустил `go mod tidy`.
- [ ] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [ ] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

## Результаты бенчмарков

Проведён сравнительный анализ производительности функции `GetDomainStat` до и после оптимизации. Тесты запускались на машине с конфигурацией:

- OS: Darwin (macOS)
- Архитектура: arm64 (Apple M1 Pro)
- Пакет: `github.com/fixme_my_friend/hw10_program_optimization`

### Сводка по времени выполнения (sec/op):

| Тест            | Старый результат (old.txt) | Новый результат (new.txt) | Изменение                 |
|-----------------|----------------------------|---------------------------|---------------------------|
| GetDomainStat-10 | 254.4ms ± 1%               | 121.6ms ± 1%              | **-52.19%** (p=0.002, n=6) |